feat: 支持歌词搜索的api

+++ src/util/format.rs
@@ -1,3 +1,3 @@
-pub(crate) fn format_duration(secs: u32) -> String {
+pub(crate) fn format_duration(secs: u64) -> String {
     format!("{:02}:{:02}", secs / 60, secs % 60)
 }
